Move images logic to new service
